move batching to separate class
